close edges to new transitions
This is the key step in resolving the top hit bug in its most general
sense. The top and right boundaries *must* both be closed, since when
the grain_hill looks for the edges, it looks for closed nodes, not
!core nodes.

With this change, the CTS model never attempts to transition a boundary
vulnerable to the fault advection methods.

The left edge is left open to permit outflow of the sediments. If it is
closed, the system will accumulate sediment in a closed box.